[bugfix][feature][documentation] Modernize the role, add a simple integration test (#11)
* QA

* add .github

* QA

* add more tests that should have been here

* create VMs

* make ansible happy

TASK [ansible-role-rabbitmq : Create erlang cookie] ****************************
[WARNING]: when statements should not include jinja2 templating delimiters
such as {{ }} or {% %}. Found: {{ not cookie.stat.exists }}

* use ubuntu 16.04, enable idempotency_test

if time permits. 14.04 would be supported

* support /etc/default and /etc/rc.conf.d

* use standard variable naming

* use standard variable naming

* correct permission is 0600, not 0400

rabbitmq changes permission of the cookie file upon startup.

remove "Fix permission" task, is is a hack

fix some typos

* add cookie file test

* add more default file tests and remove duplicated one

* remove duplicated task

the task should have been in configure-, not install-

* remove duplicated task

also, ulimit check is unnecessary in 16.04. the `Max open files` is
set to 65536 by default.

* s/rabbitmq/rabbitmq-server/

and drop ulimit which is unnecessary

* bugfix: use `-n`, do not execute the file

* update meta

* bugfix: always assign value to __variables

* adopt new permission in ubuntu

* support third-party plugins

however, you have to manually download files and place them to a local
directory.

* use same ownership across distributions

* bugfix: use quote

* bugfix: the output format is not same across versions

* initial imprementation of integration tests

* QA

* workaround inconsistent rabbitmqctl's output

the command outputs in different style. as described in issue #8, this
is nothing but a hack.

also, add an assert to make sure the plugins are registered

* support `state` in rabbitmq_plugins

* bugfix: you need `state` attribute in the module

leave a comment the task is not tested in unit test

* you cannot test the task in integration test either

* follow recent changes in the role

* connect to the build

* remove unused template

* drop templated rabbitmq.config

the old wat to manage the file is not sustainable, nor flexible. the
template would have to be modified every time you need new setting.

also, fix base directory of rabbitmq_cookie_file

* fix bug in Ubuntu.default.j2, add test for the bug

* document everything

* update rabbitmq_config, which is now mandatory

* remove one-liner hack

fixes #8

* typo